Create Jumping Frogs.java
Easy

Description:
N frogs are positioned at one end of the pond. All frogs want to reach the other end of the pond as soon as possible. The pond has some leaves arranged in a straight line. Each frog has the strength to jump exactly K leaves. For example, a frog having strength 2 will visit the leaves 2, 4, 6, ... etc. while crossing the pond.

Given the strength of each frog and the number of leaves, your task is to find the number of leaves that not be visited by any of the frogs when all frogs have reached the other end of the pond. 

Input Format:
First Line contains A Integer N

Second Line contains Integer leaves

Third Line contains Space separated Integers representing the strength of each frog

Output Format:
The number of leaves that not be visited by any of the frogs

Constraints:
1 ≤ N, leaves, frogs[i] ≤ 10^5

Public Test Cases:
Test Case 1:
Input:
3

6

1 2 4

Output:
0

Explanation:
Self Explanatory