Create Switches And Electricity Bill.java
Switches And Electricity Bill
Easy

Description:
There are n switches in a house. The number of units of electricity consumed per hour when a switch is on is given. An integer k is given and the switches' indexes divisible by k are the switches that are on. Each unit costs 12 rupees. Find the bill amount of the house for 10 hours.

Input Format:
n - number of switches in the house

arr - number of units of electricity consumed when the switch is on

k - an integer that specifies which switch is on

Output Format:
The total bill amount is for 10 hours in the house.

Constraints:
1<=n<=105

1<=arr[i]<=1000

1<=k<=n

Public Test Cases:
Test Case 1:
Input:
5

1 2 3 4 5

2

Output:
720

Explanation:
Indexes divisible by 2 are 2 and 4 one-based indexes. So 2+4=6 units per hour. Then for 10 hours is 60.Then amount =12*60 = 720 rupees

Tags:
Array Mathematical coding problems