Create Missing Chidren And The Mystery Murderer.java
Missing Chidren And The Mystery Murderer
Easy

Description:
In a town, there are n lanes and in each lane, there are a few missing kids from the past 10 days. It is believed there exists a mystery murderer who captures kids. But it is observed that there are a lot of missing kids which is not possible for a single man to do. As from the records, there is one known mystery murderer but he follows a pattern of capturing kids.

He randomly chooses one number every day and captures kids from the lanes where the lane numbers are divisible by the chosen number k. The lane number is 1 based index of the array. Find the number of children captured by that particular murderer.

Input Format:
n - number of lanes in the town

arr - number of children missing in each lane

k - number choose

Output Format:
Total number of missing children taken by the mystery murderer

Constraints:
1<=n<=105

1<=arr[i]<=1000

1<=k<=1000

Public Test Cases:
Test Case 1:
Input:
5

1 2 3 4 5

5

Output:
5

Explanation:
The lanes divisible by 5 is only the last lane and it has 5 children. So print 5.

Test Case 2:
Input:
10

1 2 3 4 5 6 7 8 9 10

4

Output:
12

Explanation:
Self Explanatory

Tags:
Array