Create Students Who Wake Up Early.java
Students Who Wake Up Early
Easy

Description:
There are n students in a hostel. And the warden comes to wake them all up to get to the classes. Some students already woke up and those who are early and students not woke up by the time the warden comes are considered lazy. But the fact is that the students who did not wake up did a lot of hard work and study the previous night and so are lacking sleep. An array is given representing the number of pages read by each student. A threshold is also given above in which the students did hard work. The students with pages equal to the threshold are considered to be waking up by the call of the Warden. Find the number of students whom the warden found to be awake

Input Format:
n- Number of students

arr - Number of pages read by each student

k- Threshold

Output Format:
Number of students who woke up before being insisted by Warden.

Constraints:
1<=n<=105

1<=arr[i]<=1000

1<=k<=1000

Public Test Cases:
Test Case 1:
Input:
5

1 2 3 3 5

2

Output:
1

Explanation:
The number of students with threshold are 4 so the students with less threshold wakes up.

Tags:
Array