Create Check If The Array Is Sorted.py
Check If The Array Is Sorted
Easy
0

0

Description:
Given an array of size n, write a program to check if the given array is sorted in (ascending / Increasing / Non-decreasing) order or not. If the array is sorted then return True, Else return False.

Two consecutive equal values are considered to be sorted.

Input Format:
n-length of the array

arr-array numbers separated by spaces

Output Format:
True if the array is sorted False if not

Constraints:
1 <= nums.length <= 104

1 <= nums[i] <= 1000

Public Test Cases:
Test Case 1:
Input:
5

1 2 3 4 5

Output:
True

Explanation:
The given array is sorted i.e. Every element in the array is smaller than or equal to its next values, So the answer is True.

Test Case 2:
Input:
5

5 4 6 7 8

Output:
False

Explanation:
The given array is Not sorted i.e. Every element in the array is not smaller than or equal to its next values, So the answer is False.

Tags:
ArraySorting