Create The Enigmatic Powers.java
The Enigmatic Powers
Easy
0

0

Description:
In the realm of numbers, there exists a special kind known as the Powers of Two. These numbers hold a unique charm, often appearing in unexpected places. But how do we recognize them amidst the vast array of numbers?



Imagine yourself as an adventurer in this numerical realm, armed with a simple question: "Is it a Power of Two?" With this question in mind, you set out on a journey to uncover the secrets of these intriguing numbers.



At each step of your journey, you encounter a new number, waiting to be explored. Your task is clear: determine whether this number possesses the magical property of being a Power of Two. Such numbers are those that can be expressed as 2x , where x is a whole number.



As you traverse through this numerical landscape, you encounter a variety of numbers, ranging from the mundane to the extraordinary. With each discovery, you gain insight into the nature of Powers of Two and their significance in mathematics.



So, let us embark on this adventure together, as we delve into the world of Powers of Two and uncover their hidden treasures.

Input Format:
First line of input contains T - number of test cases. Its followed by T lines, each line containing a single positive integer.

Output Format:
For each test case, print "True" or "False", separated by new line.

Constraints:
1 <= T <= 100

1 <= N <= 1018

Public Test Cases:
Test Case 1:
Input:
5

1

8

10

25

512

Output:
True

True

False

False

True

Explanation:
Self Explanatory